Fix avatar profile image, break out state from UI state, start asset functions
