+++ react-ui/constants/ActionTypes.js
@@ -23,9 +23,6 @@ export const ActionTypes = {
   GetCreators: 'GetCreators',
   CopyAddress: 'CopyAddress',
   CopyPrivateKey: 'CopyPrivateKey',
-  SetName: 'ChangeName',
-  SetAvatar: 'SetAvatar',
-  SetHomespace: 'SetHomespace',
   AddToLoadout: 'AddToLoadout',
   RemoveFromLoadout: 'RemoveFromLoadout',
   UploadFile: 'UploadFile',
