+++ react-ui/constants/InitialStateValues.js
@@ -6,12 +6,14 @@ export const InitialStateValues = {
   mainnetAddress: null,
   mainnetFtBalance: null,
   mainnetInventory: null,
-  avatarThumbnail: null,
   showUserDropdown: false,
   address: null,
   avatarUrl: null,
   avatarFileName: null,
   avatarPreview: null,
+  homespaceUrl: null,
+  homespaceFileName: null,
+  homespacePreview: null,
   ftu: true,
   inventory: null,
   creatorProfiles: {},
